ci: Add reusable GitHub Actions workflows for benchmarks and update configs
- Added `reusable-benchmark-publish.yml` for streamlined artifact generation and publishing across workflows.
- Updated `benchmark-iss.yml`, `benchmark-rtl.yml`, and `benchmark-lcb.yml` workflows to use the reusable workflow, reducing redundancy.
- Refactored `IssEmbenchBenchmark` to improve benchmark result handling with `outputPath`.
