+++ .github/workflows/benchmark-lcb.yml
@@ -1,10 +1,14 @@
-name: benchmark LCB
+name: Benchmark LCB
 
 on:
   workflow_dispatch:
   push:
     branches:
       - master
+    paths:
+      - 'vadl/main/vadl/lcb/**'
+      - 'vadl/test/vadl/lcb/**'
+      - 'sys/**'
 
 permissions:
   contents: read
@@ -25,73 +29,29 @@ jobs:
         run: ./gradlew assemble testClasses -PFailOnWarnings
   riscv32:
     needs: build
-    runs-on: self-hosted-x64 # The git push plugin requires linux
-    steps:
-      - uses: actions/checkout@v4
-      - uses: graalvm/setup-graalvm@v1
-        with:
-          java-version: '25'
-          distribution: 'graalvm'
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          cache: 'gradle'
-      - name: Run test
-        run: |
-          ./gradlew :vadl:test -i --tests "vadl.lcb.riscv.riscv32.EmbenchBenchmarkRiscv32Test" -Dtags.include=BenchmarkTest
-        env:
-          isCI: true
-          EMBENCH_BENCHMARK_RESULT_HOST_PATH: ${{ github.workspace }}/riscv32.csv
-          EMBENCH_BENCHMARK_RESULT_GUEST_PATH: /src/embench/result/executed_instructions_relative.csv
-      - name: Read
-        run: cat ${{ github.workspace }}/riscv32.csv
-      - name: Update csv
-        run: awk -v hash="${{ github.sha }}" -v date="${{ github.event.repository.updated_at}}" -F',' '{print $0 "," hash "," date}' ${{ github.workspace }}/riscv32.csv > ${{ github.workspace }}/${{ github.event.repository.updated_at}}-${{ github.sha }}.csv
-      - name: Read updated
-        run: cat ${{ github.workspace }}/${{ github.event.repository.updated_at }}-${{ github.sha }}.csv
-      - name: Pushes csv
-        uses: dmnemec/copy_file_to_another_repo_action@main
-        env:
-          API_TOKEN_GITHUB: ${{ secrets.ACCESS_TOKEN_RESSELPARK }}
-        with:
-          source_file: ${{ github.workspace }}/${{ github.event.repository.updated_at }}-${{ github.sha }}.csv
-          destination_repo: 'openVADL/resselpark'
-          destination_branch: 'master'
-          destination_folder: 'data/lcb/riscv32/instruction-delta/results'
-          user_email: 'ci@openvadl.org'
-          user_name: 'ci'
-          commit_message: 'data: Pushed LCB data'
+    uses: ./.github/workflows/reusable-benchmark-publish.yml
+    with:
+      runner: self-hosted-x64
+      benchmark-command: ./gradlew :vadl:test -i --tests "vadl.lcb.riscv.riscv32.EmbenchBenchmarkRiscv32Test" -Dtags.include=BenchmarkTest
+      artifact-source: riscv32.csv
+      artifact-name: ${{ github.event.repository.updated_at }}-${{ github.sha }}.csv
+      destination-folder: data/lcb/riscv32/instruction-delta/results
+      commit-message: 'data: Pushed LCB data'
+      extra-env: |
+        EMBENCH_BENCHMARK_RESULT_HOST_PATH=${{ github.workspace }}/riscv32.csv
+        EMBENCH_BENCHMARK_RESULT_GUEST_PATH=/src/embench/result/executed_instructions_relative.csv
+    secrets: inherit
   riscv64:
     needs: build
-    runs-on: self-hosted-x64 # The git push plugin requires linux
-    steps:
-      - uses: actions/checkout@v4
-      - uses: graalvm/setup-graalvm@v1
-        with:
-          java-version: '25'
-          distribution: 'graalvm'
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          cache: 'gradle'
-      - name: Run test
-        run: |
-          ./gradlew :vadl:test -i --tests "vadl.lcb.riscv.riscv64.EmbenchBenchmarkRiscv64Test" -Dtags.include=BenchmarkTest
-        env:
-          isCI: true
-          EMBENCH_BENCHMARK_RESULT_HOST_PATH: ${{ github.workspace }}/riscv64.csv
-          EMBENCH_BENCHMARK_RESULT_GUEST_PATH: /src/embench/result/executed_instructions_relative.csv
-      - name: Read
-        run: cat ${{ github.workspace }}/riscv64.csv
-      - name: Update csv
-        run: awk -v hash="${{ github.sha }}" -v date="${{ github.event.repository.updated_at}}" -F',' '{print $0 "," hash "," date}' ${{ github.workspace }}/riscv64.csv > ${{ github.workspace }}/${{ github.event.repository.updated_at}}-${{ github.sha }}.csv
-      - name: Read updated
-        run: cat ${{ github.workspace }}/${{ github.event.repository.updated_at }}-${{ github.sha }}.csv
-      - name: Pushes csv
-        uses: dmnemec/copy_file_to_another_repo_action@main
-        env:
-          API_TOKEN_GITHUB: ${{ secrets.ACCESS_TOKEN_RESSELPARK }}
-        with:
-          source_file: ${{ github.workspace }}/${{ github.event.repository.updated_at }}-${{ github.sha }}.csv
-          destination_repo: 'openVADL/resselpark'
-          destination_branch: 'master'
-          destination_folder: 'data/lcb/riscv64/instruction-delta/results'
-          user_email: 'ci@openvadl.org'
-          user_name: 'ci'
-          commit_message: 'data: Pushed LCB data'
+    uses: ./.github/workflows/reusable-benchmark-publish.yml
+    with:
+      runner: self-hosted-x64
+      benchmark-command: ./gradlew :vadl:test -i --tests "vadl.lcb.riscv.riscv64.EmbenchBenchmarkRiscv64Test" -Dtags.include=BenchmarkTest
+      artifact-source: riscv64.csv
+      artifact-name: ${{ github.event.repository.updated_at }}-${{ github.sha }}.csv
+      destination-folder: data/lcb/riscv64/instruction-delta/results
+      commit-message: 'data: Pushed LCB data'
+      extra-env: |
+        EMBENCH_BENCHMARK_RESULT_HOST_PATH=${{ github.workspace }}/riscv64.csv
+        EMBENCH_BENCHMARK_RESULT_GUEST_PATH=/src/embench/result/executed_instructions_relative.csv
+    secrets: inherit
