+++ .github/workflows/benchmark-rtl.yml
@@ -1,7 +1,14 @@
-name: benchmark RTL
+name: Benchmark RTL
 
 on:
-  workflow_dispatch: # Allows manual triggering only (for now)
+  workflow_dispatch:
+  push:
+    branches:
+      - master
+    paths:
+      - 'vadl/main/vadl/rtl/**'
+      - 'vadl/test/vadl/rtl/**'
+      - 'sys/**'
 
 permissions:
   contents: read
@@ -22,36 +29,14 @@ jobs:
         run: ./gradlew assemble testClasses -PFailOnWarnings
   riscv:
     needs: build
-    runs-on: self-hosted-x64 # The git push plugin requires linux
-    steps:
-      - uses: actions/checkout@v4
-      - uses: graalvm/setup-graalvm@v1
-        with:
-          java-version: '25'
-          distribution: 'graalvm'
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          cache: 'gradle'
-      - name: Run test
-        run: |
-          ./gradlew test -i --tests "vadl.rtl.riscv.RtlRiscVBenchmarkTest" -Dtags.include=BenchmarkTest
-        env:
-          isCI: true
-          RTL_DEC_BENCHMARK_RESULT_HOST_PATH: ${{ github.workspace }}/riscv_rtl_dec_benchmarks.csv
-      - name: Read
-        run: cat ${{ github.workspace }}/riscv_rtl_dec_benchmarks.csv
-      - name: Update csv
-        run: awk -v hash="${{ github.sha }}" -v date="${{ github.event.repository.updated_at}}" -F',' 'NR==1{print $0 ",commit_sha,repo_updated_at"} NR>1{print $0 "," hash "," date}' ${{ github.workspace }}/riscv_rtl_dec_benchmarks.csv > ${{ github.workspace }}/${{ github.event.repository.updated_at}}-${{ github.sha }}.csv
-      - name: Read updated
-        run: cat ${{ github.workspace }}/${{ github.event.repository.updated_at }}-${{ github.sha }}.csv
-      - name: Pushes csv
-        uses: dmnemec/copy_file_to_another_repo_action@main
-        env:
-          API_TOKEN_GITHUB: ${{ secrets.ACCESS_TOKEN_RESSELPARK }}
-        with:
-          source_file: ${{ github.workspace }}/${{ github.event.repository.updated_at }}-${{ github.sha }}.csv
-          destination_repo: 'openVADL/resselpark'
-          destination_branch: 'master'
-          destination_folder: 'data/rtl/riscv'
-          user_email: 'ci@openvadl.org'
-          user_name: 'ci'
-          commit_message: 'data: Pushed RTL data'
+    uses: ./.github/workflows/reusable-benchmark-publish.yml
+    with:
+      runner: self-hosted-x64
+      benchmark-command: ./gradlew test -i --tests "vadl.rtl.riscv.RtlRiscVBenchmarkTest" -Dtags.include=BenchmarkTest
+      artifact-source: riscv_rtl_dec_benchmarks.csv
+      artifact-name: ${{ github.event.repository.updated_at }}-${{ github.sha }}.csv
+      destination-folder: data/rtl/riscv/benchmark
+      commit-message: 'data: Pushed RTL data'
+      extra-env: |
+        RTL_DEC_BENCHMARK_RESULT_HOST_PATH=${{ github.workspace }}/riscv_rtl_dec_benchmarks.csv
+    secrets: inherit
